提交资源文件解析

+++ src/com/erlin/parse/androidArsc/type/ResTableType.java
@@ -8,7 +8,7 @@ public class ResTableType {
     public ResChunkHeader resChunkHeader;
     public byte id;
     public byte res0;
-    public byte res1;
+    public short res1;
     public int entryCount;
     public int entriesStart;
     public ResTableConfig resTableConfig;
